requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.2
astroid==3.2.4
async-timeout==5.0.1
attrs==24.3.0
autopep8==2.3.1
awscli==1.36.31
backports.zoneinfo==0.2.1
black==24.8.0
blinker==1.8.2
botocore==1.35.90
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
dateparser==1.2.0
dill==0.3.9
docutils==0.16
exceptiongroup==1.2.2
fastapi==0.115.6
fonttools==4.55.3
frozendict==2.4.6
frozenlist==1.5.0
h11==0.14.0
html5lib==1.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
kiwisolver==1.4.7
lxml==5.3.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mccabe==0.7.0
multidict==6.1.0
multitasking==0.0.11
mypy-extensions==1.0.0
mysql-connector-python==9.0.0
numpy==1.24.4
packaging==24.2
pandas==2.0.3
pathspec==0.12.1
peewee==3.17.8
pillow==10.4.0
platformdirs==4.3.6
propcache==0.2.0
pyasn1==0.6.1
pycodestyle==2.12.1
pycryptodome==3.21.0
pydantic==2.10.4
pydantic_core==2.27.2
pylint==3.2.7
PyMySQL==1.1.1
pyparsing==3.1.4
python-binance==1.0.27
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.7.2
s3transfer==0.10.4
schedule==1.2.2
scipy==1.10.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
starlette==0.41.3
tomli==2.2.1
tomlkit==0.13.2
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==1.26.20
uvicorn==0.33.0
webencodings==0.5.1
websockets==13.1
Werkzeug==3.0.6
yarl==1.15.2
zipp==3.20.2
